Alan branch (#5)
* fixed bug with variable misnamed in Signup.php

* created a search contact feature and functions to create and verify JWT

* implemented hashing for password

* Delete contact created

Created Delete contact file, token of user and contact id to delete is needed as params. Contact is also returned in returnWithContactInfo function

* Duplicate Login error and deleteUser function

Not sure if deleteUser.php is needed

* Delete contact created

Created Delete contact file, token of user and contact id to delete is needed as params. Contact is also returned in returnWithContactInfo function

* Duplicate Login error and deleteUser function

Not sure if deleteUser.php is needed

* created a search contact feature and functions to create and verify JWT

* implemented hashing for password

* Delete contact created

Created Delete contact file, token of user and contact id to delete is needed as params. Contact is also returned in returnWithContactInfo function

* Create edit contact

* Edit Contact php Finished

* Login returns more User Info w/ Token

* composer

---------

Co-authored-by: Angelo <angelowhite871@gmail.com>